♻️ Split homepage into sections
